refactoring bigrams stuff and adding more tests
