+++ setup.cfg
@@ -1,5 +1,6 @@
 [flake8]
 extend-ignore = E203,E501
+exclude = dist
 
 [mypy]
 follow_imports = silent
@@ -11,6 +12,7 @@ check_untyped_defs = True
 disallow_untyped_defs = True
 namespace_packages = True
 mypy_path = $MYPY_CONFIG_FILE_DIR/stubs
+exclude = dist
 
 [mypy-tests.*]
 ignore_missing_imports = True
